데이터베이스를 연동한 로그인후 진행하는 가위바위보

+++ src/D20240930/Board/DTO/CommentDTO.java
@@ -13,7 +13,7 @@ public class CommentDTO {
   private java.sql.Timestamp updtDt;
 
 
-  public long getCNo() {
+  public int getCNo() {
     return cNo;
   }
